celery-sample/requirements.txt
amqp==5.0.9
anyio==3.5.0
asgiref==3.5.0
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.73.0
flower==1.0.0
h11==0.13.0
httptools==0.3.0
humanize==3.13.1
idna==3.3
kombu==5.2.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.1.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==1.2.3
tornado==6.1
typing_extensions==4.0.1
uvicorn==0.17.4
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
wrapt==1.13.3
